+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.18.8 Position-only cover 2021-06.30
+
+### Devices
+
+- Cover: enable `set_up` and `set_down` with `group_address_position` only (without `group_address_long`).
+
 ## 0.18.7 RawValue 2021-06-18
 
 ### Devices